python-requirements.txt
alabaster==0.7.10
Babel==2.5.1
certifi==2017.7.27.1
chardet==3.0.4
chartkick==0.5.0
Django==2.0
django-prettyjson==0.3.0
djangorestframework==3.7.7
docutils==0.14
idna==2.6
imagesize==0.7.1
Jinja2==2.9.6
MarkupSafe==1.0
pkg-resources==0.0.0
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.4
sphinxcontrib-websupport==1.0.1
standardjson==0.3.1
urllib3==1.22
